zephyr: Add console snippet
Adds a new zephyr build system snippet that enables the iiod server over
the console UART. This demonstrates adding iiod to the in-tree hello
world sample application without modifying the application:

$ west build -b apard32690/max32690/m4 \
    ~/zephyrproject/zephyr/samples/hello_world/ \
    -S iiod-console

$ west build -b frdm_k64f \
    ~/zephyrproject/zephyr/samples/hello_world/ \
    -S iiod-console

$ west build -b adi_sdp_k1 \
    ~/zephyrproject/zephyr/samples/hello_world/ \
    -S iiod-console

The console and iiod are directed to the same uart device, therefore is
is not possible to communicate with both simultaneously. libiio
effectively takes over the uart device from the console.

Signed-off-by: Maureen Helm <maureen.helm@analog.com>